add notification from company to user

+++ cleaning_service/final_project/core/models/profiles.py
@@ -28,7 +28,7 @@ class User(AbstractBaseUser, PermissionsMixin):  # Base user`s model overriding
     services = models.ManyToManyField(Service)
     phone = models.CharField(verbose_name="User`s phone", max_length=100, null=False, default="0")
     role = models.ForeignKey(to=UserRole, on_delete=models.CASCADE, null=False)
-    rating = models.PositiveSmallIntegerField(verbose_name="Star-rating", default=0, null=False, validators=(
+    rating = models.FloatField(verbose_name="Star-rating", default=0, null=False, validators=(
         MaxValueValidator(5), MinValueValidator(0)))
 
     # User`s rights
